net/smc: move smc_sock related structure definition
This patch only try to move the definition of smc_sock and its
related structure, from et/smc/smc.h to include/net/smc/smc.h.
In that way can ebpf generate the BTF ID corresponding to our
structure.

Of course, we can also choose to hide the structure and only to
expose an intermediate structure, but it requires an additional
transformation. If we need to obtain some information frequently, this
may cause some performance problems.
